moving time division into paje

+++ src/prvreader/headers/interpretercomponent.h
@@ -20,10 +20,10 @@ namespace prvreader{
     class InterpreterComponent
     {
     public:
-        virtual void pushEvents(int cpu, int app, int task, int thread, double timestamp, map<int, string>* events, long lineNumber)=0;
-        virtual void pushState(int cpu, int app, int task, int thread, double startTimestamp, double endTimestamp, string value, long lineNumber)=0;
+        virtual void pushEvents(int cpu, int app, int task, int thread, long timestamp, map<int, string>* events, long lineNumber)=0;
+        virtual void pushState(int cpu, int app, int task, int thread, long startTimestamp, long endTimestamp, string value, long lineNumber)=0;
         virtual void pushCommunications(int cpu1, int app1, int task1, int thread1, int cpu2, int app2, int task2, int thread2,
-                                        double startTimestampSW, double startTimestampHW, double endTimestampSW, double endTimestampHW, string value, long lineNumber)=0;
+                                        long startTimestampSW, long startTimestampHW, long endTimestampSW, long endTimestampHW, string value, long lineNumber)=0;
         virtual void initialize()=0;
         virtual void finalize()=0;
